requirements.txt
action==1.4.4
allure-behave==2.8.40
allure-pytest==2.8.40
allure-python-commons==2.8.40
anaconda==0.0.1.1
appdirs==1.4.4
Appium-Python-Client==2.0.0b2
asgiref==3.3.4
astropy==4.2.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
behave==1.2.6
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0a1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==3.4.7
distlib==0.3.1
Django==3.2
djangorestframework==3.12.4
dplython==0.0.7
driver==0.1.0
extras==1.0.0
filelock==3.0.12
fixtures==3.0.0
fuzzywuzzy==0.18.0
gevent==21.1.2
greenlet==1.0.0
h11==0.12.0
hs-test-python==4.0.0
idna==3.1
ignore==0.1.4
iniconfig==1.1.1
joblib==1.0.1
lettuce==0.2.23
Library==0.0.0
linecache2==1.0.0
mock==4.0.3
more-itertools==8.7.0
mysqlclient==2.0.3
nose==1.3.7
numpy==1.20.2
outcome==1.1.0
packaging==20.9
pandas==1.2.4
parse==1.19.0
parse-type==0.5.2
pbr==5.6.0
Pillow==8.2.0
pluggy==1.0.0.dev0
psutil==5.8.0
py==1.10.0
py-zap==1.2.2
pycparser==2.20
pyerfa==2.0.0
pygame==2.0.0
pyOpenSSL==20.0.1
pyparsing==3.0.0b2
pyresttest==1.5.0
pyTelegramBotAPI==3.7.7
pytest==6.2.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
python-owasp-zap-v2.4==0.0.18
python-subunit==1.4.0
pytidy==0.4.0
pytidylib==0.3.2
pytz==2021.1
readpy==0.0.1.dev2
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
selenium==4.0.0b2.post1
Selenium-Screenshot==1.7.0z
simple-draw==2.7.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.3.0
soupsieve==2.2.1
sqlparse==0.4.1
style==1.1.6
sure==1.4.11
swampy==3.0.1
telebot==0.0.4
testtools==2.4.0
threadpoolctl==2.1.0
tidypython==0.0.1.dev3
toml==0.10.2
traceback2==1.4.0
trio==0.18.0
trio-websocket==0.9.2
unittest2==1.1.0
update==0.0.1
urllib3==1.26.4
virtualenv==20.4.3
watchdog==2.0.2
wcwidth==0.2.5
webdriver-manager==3.3.0
wsproto==1.0.0
zap==2.1
zope.event==4.5.0
zope.interface==5.3.0
